+++ src/main/java/it/lisik/itunesreceiptvalidator/deserializer/ReceiptMixin.java
@@ -9,10 +9,10 @@ import java.util.Set;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 abstract class ReceiptMixin {
-    ReceiptMixin(@JsonProperty("receipt_type") String receiptType, @JsonProperty("adam_id") Integer adamId,
-                 @JsonProperty("app_item_id") Integer appItemId, @JsonProperty("bundle_id") String bundleId,
+    ReceiptMixin(@JsonProperty("receipt_type") String receiptType, @JsonProperty("adam_id") Long adamId,
+                 @JsonProperty("app_item_id") Long appItemId, @JsonProperty("bundle_id") String bundleId,
                  @JsonProperty("application_version") String applicationVersion,
-                 @JsonProperty("download_id") Integer downloadId,
+                 @JsonProperty("download_id") Long downloadId,
                  @JsonProperty("version_external_identifier") Integer versionExternalIdentifier,
                  @JsonProperty("receipt_creation_date") DateTime receiptCreationDate,
                  @JsonProperty("request_date") DateTime requestDate,
